requirements.txt
tqdm
numpy
pillow
omegaconf
matplotlib
torch~=2.4.1
torchvision~=0.19.1
accelerate~=1.0.1
tensorboard~=2.18.0
git+https://github.com/toshas/torch-fidelity.git@master#egg=torch-fidelity
